Fixing the Loading Problem
I was having issue with how to make the loading screen load levels in the intended order, but I managed to fix it. What I did was create a public static variable in the Level Manager Script, which is added to every level, and made it x. Every time the Player collides with the triangle at the end of the level, coded to bring up the loading screen, it increases the x value by one. This x value tells Unity to load the scene with that BuildIndex number.